requirements.txt
# Web 框架与 ASGI 服务
fastapi>=0.115.0
uvicorn[standard]>=0.30.0,<0.32.0

# 数据校验与配置
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0

# HTTP 客户端
httpx>=0.28.0

# 数据库 (SQLite 异步)
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# AWS event stream 解析（Kiro 通道）
botocore>=1.34.0

# 日志与工具库
loguru>=0.7.2,<1.0.0
shortuuid>=1.0.11,<2.0.0

# 环境变量
python-dotenv>=1.0.0

# YAML 配置解析
pyyaml>=6.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.24.0
hypothesis>=6.100.0
